+++ src/fonts/BravoSCT/fonts.css
@@ -1,12 +1,9 @@
 @font-face {
-    font-family: 'BravoSCT';
-    src: url('../fonts/BravoSCT/BravoSCT.eot');
-    src: url('../fonts/BravoSCT/BravoSCT.eot?#iefix') format('embedded-opentype'),
-        url('../fonts/BravoSCT/BravoSCT.woff2') format('woff2'),
-        url('../fonts/BravoSCT/BravoSCT.woff') format('woff'),
-        url('../fonts/BravoSCT/BravoSCT.ttf') format('truetype'),
-        url('../fonts/BravoSCT/BravoSCT.svg#BravoSCT') format('svg');
-    font-weight: normal;
-    font-style: normal;
+  font-family: 'BravoSCT';
+  src: url('../fonts/BravoSCT/BravoSCT.woff2') format('woff2'),
+  url('../fonts/BravoSCT/BravoSCT.woff') format('woff'),
+  url('../fonts/BravoSCT/BravoSCT.ttf') format('truetype');
+  font-weight: normal;
+  font-style: normal;
 }
