+++ src/main.js
@@ -1,13 +1,13 @@
 import './libs/other-libs/bootstrap-grid.css'
 import './libs/other-libs/reset.css'
 
-import './js/goda-page'
+import './js/years-page'
 import './js/history-page'
 import './js/home-page'
 import './js/victims-page'
 
 import './styles/link.scss'
-import './styles/goda-page.scss'
+import './styles/years-page.scss'
 import './styles/history-page.scss'
 import './styles/home-page.scss'
 import './styles/victims-page.scss'
