+++ src/pages/index.twig
@@ -11,7 +11,7 @@
       <div class="home__block">
         <img src="img/home-block-1.jpg" alt="">
         <div class="home__block_link">
-          <a href="goda-page.html" class="link">
+          <a href="years-page.html" class="link">
             система
           </a>
         </div>
